testing/web-platform/tests/atomics/window-and-any-worker/wake/wake-all-on-loc/wake-all-on-loc.html

<!doctype html>
<meta charset=utf-8>
<title></title>

<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>

<body>
  <iframe id="myiframe" style="display: none" src="../empty.html" ></iframe>
</body>

<script>

// Copyright (C) 2017 Mozilla Corporation.  All rights reserved.
// This code is governed by the BSD license found in the LICENSE file.

/*---
description: >
  Test that Atomics.wake wakes all waiters on a location, but does not
  wake waiters on other locations.
---*/

window.addEventListener("load", function() {
  async_test(function(t) {
    let rs = [];
    let workers = [];
    let nworkers = 4;

    // Create three workers and make them wait on zero.
    for (let i = 0; i < nworkers-1; i++) {
      let w = new Worker('../wait-on-zero.js');
      w.onmessage = function(e) {
        rs.push(e.data);
      }
      workers.push(w);
    }
    // Make fourth worker and make it wait on one.
    let w = new Worker('../wait-on-one.js');
    w.onmessage = function(e) {
      rs.push(e.data);
    }
    workers.push(w);
    // Send message to all workers.
    let ia = new Int32Array(new SharedArrayBuffer(2*Int32Array.BYTES_PER_ELEMENT));
    for (let i = 0; i < nworkers; i++) {
      workers[i].postMessage([ia.buffer]);
    }
    // Wake workers waiting on zero.
    setTimeout(() => assert_equals(3, Atomics.wake(ia, 0)), 500);
    // Check workers on zero woke but worker on one timed-out.
    setTimeout(this.step_func_done(function() {
      rs.sort();
      assert_equals(rs[0], "ok");
      assert_equals(rs[1], "ok");
      assert_equals(rs[2], "ok");
      assert_equals(rs[3], "timed-out");
    }), 2000);
  }, "atomics-dedicated-worker-wake-all-on-loc");
  async_test(function(t) {
    let worker = new Worker('wake-all-on-loc-worker.js');
    worker.onmessage = this.step_func_done(function(e) {
      let rs = e.data[0];
      rs.sort();
      console.log("worker");
      for (let i = 0; i < rs.length; i++) {
        console.log("rs: " + rs[i]);
      }
      assert_equals(rs[0], "ok");
      assert_equals(rs[1], "ok");
      assert_equals(rs[2], "ok");
      assert_equals(rs[3], "timed-out");
    });
    worker.postMessage([]);
  }, "atomics-dedicated-worker-wake-all-on-loc-from-worker");
  async_test(function(t) {
    let worker = new SharedWorker('wake-all-on-loc-shared-worker.js');
    worker.port.start();
    worker.port.onmessage = this.step_func_done(function(e) {
      let rs = e.data[0];
      rs.sort();
      console.log('rs.length: ' + rs.length);
      assert_equals(rs[0], "ok");
      assert_equals(rs[1], "ok");
      assert_equals(rs[2], "ok");
      assert_equals(rs[3], "timed-out");
    });
    worker.port.postMessage([]);
  }, "atomics-dedicated-worker-wake-all-on-loc-from-shared-worker");
  async_test(function(t) {
    let iframe = document.getElementById('myiframe');
    let contentWindow = iframe.contentWindow;
    let test = this;
    contentWindow.onmessage = function(e) {
      let rs = [];
      let workers = [];
      let nworkers = 4;

      // Create three workers and make them wait on zero.
      for (let i = 0; i < nworkers-1; i++) {
        let w = new Worker('../wait-on-zero.js');
        w.onmessage = function(e) {
          rs.push(e.data);
        }
        workers.push(w);
      }
      // Make fourth worker and make it wait on one.
      let w = new Worker('../wait-on-one.js');
      w.onmessage = function(e) {
        rs.push(e.data);
      }
      workers.push(w);
      // Send message to all workers.
      let ia = new Int32Array(new SharedArrayBuffer(2*Int32Array.BYTES_PER_ELEMENT));
      for (let i = 0; i < nworkers; i++) {
        workers[i].postMessage([ia.buffer]);
      }
      // Wake workers waiting on zero.
      setTimeout(() => assert_equals(3, Atomics.wake(ia, 0)), 500);
      // Check workers on zero woke but worker on one timed-out.
      setTimeout(test.step_func_done(function() {
        rs.sort();
        assert_equals(rs[0], "ok");
        assert_equals(rs[1], "ok");
        assert_equals(rs[2], "ok");
        assert_equals(rs[3], "timed-out");
      }), 2000);
    }
    contentWindow.postMessage([], '*');
  }, "atomics-dedicated-worker-wake-all-on-loc-from-iframe");
});
</script>
